requirements.txt
pillow==12.1.0
git+https://github.com/hbldh/hitherdither
